Limit target-syslibs and target-utils to SDE >9.9.0

+++ stratum/hal/bin/barefoot/build-bf-sde.sh
@@ -175,30 +175,32 @@ popd
 
 echo "BF SDE build complete."
 
-echo "Build and install target-syslibs and target-utils."
-TARGET_SYSLIBS_TMP=$(mktemp -d)
-git clone https://github.com/p4lang/target-syslibs.git "$TARGET_SYSLIBS_TMP"
-pushd "$TARGET_SYSLIBS_TMP"
-git checkout "$TARGET_SYSLIBS_COMMIT"
-git submodule update --init
-mkdir -p build
-cd build
-cmake -DCMAKE_INSTALL_PREFIX="$SDE_INSTALL" ..
-make install -j "$JOBS"
-popd
-rm -rf "$TARGET_SYSLIBS_TMP"
+if [[ $(numeric_version "$SDE_VERSION") -ge $(numeric_version "9.9.0") ]]; then
+    echo "Build and install target-syslibs and target-utils."
+    TARGET_SYSLIBS_TMP=$(mktemp -d)
+    git clone https://github.com/p4lang/target-syslibs.git "$TARGET_SYSLIBS_TMP"
+    pushd "$TARGET_SYSLIBS_TMP"
+    git checkout "$TARGET_SYSLIBS_COMMIT"
+    git submodule update --init
+    mkdir -p build
+    cd build
+    cmake -DCMAKE_INSTALL_PREFIX="$SDE_INSTALL" ..
+    make install -j "$JOBS"
+    popd
+    rm -rf "$TARGET_SYSLIBS_TMP"
 
-TARGET_UTILS_TMP=$(mktemp -d)
-git clone https://github.com/p4lang/target-utils.git "$TARGET_UTILS_TMP"
-pushd "$TARGET_UTILS_TMP"
-git checkout ${TARGET_UTILS_COMMIT}
-git submodule update --init --recursive
-mkdir -p build
-cd build
-cmake -DCMAKE_INSTALL_PREFIX="$SDE_INSTALL" -DSTANDALONE=1 ..
-make install -j "$JOBS"
-popd
-rm -rf "$TARGET_UTILS_TMP"
+    TARGET_UTILS_TMP=$(mktemp -d)
+    git clone https://github.com/p4lang/target-utils.git "$TARGET_UTILS_TMP"
+    pushd "$TARGET_UTILS_TMP"
+    git checkout ${TARGET_UTILS_COMMIT}
+    git submodule update --init --recursive
+    mkdir -p build
+    cd build
+    cmake -DCMAKE_INSTALL_PREFIX="$SDE_INSTALL" -DSTANDALONE=1 ..
+    make install -j "$JOBS"
+    popd
+    rm -rf "$TARGET_UTILS_TMP"
+fi
 
 # Strip shared libraries and fix permissions
 find $SDE_INSTALL -name "*\.so*" -a -type f | xargs -n1 chmod u+w
